+++ deezy/payloads/shared.py
@@ -18,7 +18,6 @@ class CorePayload:
     bitrate: int | None
     temp_dir: Path | None
     delay: str | None
-    parse_elementary_delay: bool
     keep_temp: bool
     reuse_temp_files: bool
     file_output: Path | None